Removed pointless index file
